重构(v0.21): types.ts 按消费域拆分到 src/types/ + ReportMeta 加 schemaVersion
859 行的 src/types.ts 拆到 6 个分域文件(shared/executor/judge/eval/report/
storage),原 types.ts 保留为 1 行 facade 维持现有 import 路径不变(共 95+ 处
'../types.js'). 顺手在 ReportMeta 加 schemaVersion?: number 字段为 Phase C
gap 字段重命名 + 旧报告 fallback 做准备(无字段 = v0).

- src/types/shared.ts: Lang
- src/types/executor.ts: ExecResult / ExecutorFn / ToolCallInfo / Mcp* (9 type)
- src/types/judge.ts: JudgeConfig / EnsembleJudgeResult / JudgeAgreement
  / Assertion* / DimensionResult / LayeredScores / GradeResult (8 type)
- src/types/eval.ts: Sample / Artifact / Evaluation* / Progress* / Task (27 type)
- src/types/report.ts: Report / ReportMeta / VariantSummary / Variance*
  / GapReport / SaturationData (24 type)
- src/types/storage.ts: ReportStore / JobStore
- src/types/index.ts: 全量 re-export
- src/types.ts: 改为 'export * from ./types/index.js' facade

Phase A.1 + A.2 of v0.21 测评闭环产品化攻坚 plan.
