+++ src/types/shared.ts
@@ -1,0 +1,2 @@
+// For renderer functions that accept partial report-like objects
+export type Lang = 'zh' | 'en';
